requirements.txt
starlette==0.45.1
uvicorn[standard]==0.34.0
requests==2.32.3
dadaptation==3.2
pyngrok==7.2.2
pycloudflared==0.2.0
scipy==1.15.1
came-pytorch==0.1.3
pytorch_optimizer==3.4.1
wheel==0.45.1
triton==3.2.0; platform_system == "Linux"
prodigy-plus-schedule-free==1.9.0
